Add files via upload
A dataset has been compiled for a cohort of patients all afflicted with the same illness. Throughout their treatment, each patient exhibited a response to one of five drugs: Drug A, Drug B, Drug C, Drug X, and Drug Y.
This dataset allows for the development of a predictive model to determine which drug would be suitable for a future patient with the same disease. The dataset encompasses various patient attributes, including age, sex, blood pressure, cholesterol levels, and the sodium-to-potassium ratio in the blood. The objective is to identify the drug to which each patient positively responded.